+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
+- Argo Workflow resource
 - ECS Service wait function based on deployment status instead of task state
 
 ## Fixed